requirements.txt
beautifulsoup4==4.5.1
Django==1.10
futures==3.0.5
python-dateutil==2.5.3
pytz==2016.6.1
six==1.10.0
